Include last changes of master

+++ pipeline_TST_functions.sh
@@ -97,6 +97,7 @@ outputConfigurationVariablesInformation() {
         "·Approximately size of the field:$sizeOfOurFieldDegrees:[deg]"
         "·Size of the coadd:$coaddSizePx:[px]"
         " "
+        "·Width of the normalisation ring:$ringWidth:[px]"
         "·Common normalisation ring:$USE_COMMON_RING"
         "  If so, the file with the ring specification is:$commonRingDefinitionFile"
         ""
@@ -194,6 +195,7 @@ checkIfAllVariablesAreSet() {
                 numberOfFWHMForPhotometry \
                 surveyForPhotometry \
                 starMagnitudeThresholdToReject_gBand \
+                ringWidth \
                 USE_COMMON_RING \
                 commonRingDefinitionFile \
                 keyWordToDecideRing
@@ -424,8 +426,8 @@ getSkewKurtoValueInsideRing(){
     if [ "$useCommonRing" = true ]; then
             # Case when we have one common normalisation ring
             #astarithmetic $i -h1 $commonRing -h1 0 eq nan where -q -o ring_masked.fits
-            skew=$(python $pythonScriptsPath/get_skewness_kurtosis.py $i SKEWNESS $commonRing)
-            kurto=$(python $pythonScriptsPath/get_skewness_kurtosis.py $i KURTOSIS $commonRing)
+            skew=$(python3 $pythonScriptsPath/get_skewness_kurtosis.py $i SKEWNESS $commonRing)
+            kurto=$(python3 $pythonScriptsPath/get_skewness_kurtosis.py $i KURTOSIS $commonRing)
             rm ring_masked.fits
     else
         # Case when we do NOT have one common normalisation ring
@@ -438,13 +440,13 @@ getSkewKurtoValueInsideRing(){
 
         if (( $(echo "$variableToDecideRingToNormalise >= $firstRingLowerBound" | bc -l) )) && (( $(echo "$variableToDecideRingToNormalise <= $firstRingUpperBound" | bc -l) )); then
             #astarithmetic $i -h1 $doubleRing_first -h1 0 eq nan where -q -o ring_masked.fits
-            skew=$(python $pythonScriptsPath/get_skewness_kurtosis.py $i SKEWNESS $doubleRing_first)
-            kurto=$(python $pythonScriptsPath/get_skewness_kurtosis.py $i KURTOSIS $doubleRing_first)
+            skew=$(python3 $pythonScriptsPath/get_skewness_kurtosis.py $i SKEWNESS $doubleRing_first)
+            kurto=$(python3 $pythonScriptsPath/get_skewness_kurtosis.py $i KURTOSIS $doubleRing_first)
             rm ring_masked.fits
         elif (( $(echo "$variableToDecideRingToNormalise >= $secondRingLowerBound" | bc -l) )) && (( $(echo "$variableToDecideRingToNormalise <= $secondRingUpperBound" | bc -l) )); then
             #astarithmetic $i -h1 $doubleRing_second -h1 0 eq nan where -q -o ring_masked.fits
-            skew=$(python $pythonScriptsPath/get_skewness_kurtosis.py $i SKEWNESS $doubleRing_second)
-            kurto=$(python $pythonScriptsPath/get_skewness_kurtosis.py $i KURTOSIS $doubleRing_second)
+            skew=$(python3 $pythonScriptsPath/get_skewness_kurtosis.py $i SKEWNESS $doubleRing_second)
+            kurto=$(python3 $pythonScriptsPath/get_skewness_kurtosis.py $i KURTOSIS $doubleRing_second)
             rm ring_masked.fits
         else
             errorNumber=5
@@ -584,7 +586,6 @@ runNoiseChiselOnFrame() {
 
     imageToUse=$inputFileDir/$baseName
     output=$outputDir/$baseName
-    echo astnoisechisel $imageToUse $noiseChiselParams -o $output
     astnoisechisel $imageToUse $noiseChiselParams -o $output
 }
 export -f runNoiseChiselOnFrame
@@ -651,10 +652,13 @@ removeBadFramesFromReduction() {
 
     filePath=$badFilesWarningDir/$badFilesWarningFile
 
+
     while IFS= read -r file_name; do
         file_name=$(basename "$file_name")
-        fileName="${file_name%.*}".fits
-        mv $sourceToRemoveFiles/$fileName $destinationDir/$fileName
+        fileName="entirecamera_${file_name%.*}".fits
+        if [ -f $sourceToRemoveFiles/$fileName ]; then
+            mv $sourceToRemoveFiles/$fileName $destinationDir/$fileName
+        fi
     done < "$filePath"
 }
 export -f removeBadFramesFromReduction
@@ -674,6 +678,7 @@ computeSkyForFrame(){
     keyWordThreshold=${11}
     keyWordValueForFirstRing=${12}
     keyWordValueForSecondRing=${13}
+    ringWidth=${14}
 
     i=$entiredir/$1
 
@@ -720,7 +725,7 @@ computeSkyForFrame(){
 
             ringRadius=$( awk '{print $5}' $ringDir/ring.txt )
             echo "1 $half_naxis1 $half_naxis2 6 $ringRadius 1 1 1 1 1" > $ringDir/$tmpRingDefinition
-            astmkprof --background=$imageToUse  -h1 --mforflatpix --mode=img --type=uint8 --circumwidth=200 --clearcanvas -o $ringDir/$tmpRingFits $ringDir/$tmpRingDefinition
+            astmkprof --background=$imageToUse  -h1 --mforflatpix --mode=img --type=uint8 --circumwidth=$ringWidth --clearcanvas -o $ringDir/$tmpRingFits $ringDir/$tmpRingDefinition
 
             me=$(getMedianValueInsideRing $imageToUse  $ringDir/$tmpRingFits "" "" true $keyWordToDecideRing $keyWordThreshold $keyWordValueForFirstRing $keyWordValueForSecondRing)
             std=$(getStdValueInsideRing $imageToUse $ringDir/$tmpRingFits "" "" true $keyWordToDecideRing $keyWordThreshold $keyWordValueForFirstRing $keyWordValueForSecondRing)
@@ -728,7 +733,7 @@ computeSkyForFrame(){
 
             echo "$base $me $std $skew $kurto" > $noiseskydir/$out
 
-            # rm $ringDir/$tmpRingDefinition
+            rm $ringDir/$tmpRingDefinition
             rm $ringDir/$tmpRingFits
 
         elif [ "$constantSkyMethod" = "noisechisel" ]; then
@@ -784,20 +789,22 @@ computeSky() {
     keyWordThreshold=${11}
     keyWordValueForFirstRing=${12}
     keyWordValueForSecondRing=${13}
+    ringWidth=${14}
     
     if ! [ -d $noiseskydir ]; then mkdir $noiseskydir; fi
     if [ -f $noiseskydone ]; then
         echo -e "\n\tScience images are 'noisechiseled' for constant sky substraction for extension $h\n"
     else
         framesToComputeSky=()
-        for a in $(seq 1 $totalNumberOfFrames); do
-            base="entirecamera_"$a.fits
+        for a in $( ls $framesToUseDir/*.fits ); do
+            base=$( basename $a )
             framesToComputeSky+=("$base")
         done
-        printf "%s\n" "${framesToComputeSky[@]}" | parallel -j "$num_cpus" computeSkyForFrame {} $framesToUseDir $noiseskydir $constantSky $constantSkyMethod $polyDegree $inputImagesAreMasked $ringDir $useCommonRing $keyWordToDecideRing $keyWordThreshold $keyWordValueForFirstRing $keyWordValueForSecondRing
+        printf "%s\n" "${framesToComputeSky[@]}" | parallel -j "$num_cpus" computeSkyForFrame {} $framesToUseDir $noiseskydir $constantSky $constantSkyMethod $polyDegree $inputImagesAreMasked $ringDir $useCommonRing $keyWordToDecideRing $keyWordThreshold $keyWordValueForFirstRing $keyWordValueForSecondRing $ringWidth
         echo done > $noiseskydone
     fi
 }
+export -f computeSky
 
 subtractSkyForFrame() {
     a=$1
@@ -853,7 +860,7 @@ subtractSky() {
 export -f subtractSky
 
 # Functions for decals data
-# The function that is to be used (the 'public' function using OOP terminology) is 'prepareDecalsDataForPhotometricCalibration'
+# The function that is to be used (the 'public' function using OOP terminology) is 'prepareSurveyDataForPhotometricCalibration'
 getBricksWhichCorrespondToFrame() {
     frame=$1
     frameBrickMapFile=$2
@@ -907,6 +914,7 @@ downloadData() {
     gaiaCatalogue=${11}
     starThresholdForRejectingBricks=${12}
     survey=${13}
+    
     echo -e "\n·Downloading ${survey} bricks"
     donwloadMosaicDone=$surveyImagesDir/done_downloads.txt
     if ! [ -d $mosaicDir ]; then mkdir $mosaicDir; fi
@@ -1288,6 +1296,11 @@ performAperturePhotometryToSingleBrick() {
     # photometryOnImage_noisechisel $brick $outputCat $automaticCatalogue $brickImage $r_decals_pix $outputCat/$brick.cat \
     #                             22.5 $dataHdu $raColumnName $decColumnName
 
+    # This code is only for checking the same aperture as the sloan SDSS spectrograph fiber
+    # r_decals_pix=4
+    # echo "Realizando fotometría en brick $brick con apertura: $r_decals_pix"
+
+
     columnWithXCoordForDecalsPx=0 # These numbers come from how the catalogue of the matches stars is built. This is not very clear right now, should be improved
     columnWithYCoordForDecalsPx=1
     columnWithXCoordForDecalsWCS=2
@@ -1336,7 +1349,7 @@ performAperturePhotometryToBricks() {
 export -f performAperturePhotometryToBricks
 
 
-prepareDecalsDataForPhotometricCalibration() {
+prepareSurveyDataForPhotometricCalibration() {
     referenceImagesForMosaic=$1
     surveyImagesDir=$2
     filter=$3
@@ -1416,7 +1429,7 @@ prepareDecalsDataForPhotometricCalibration() {
     fi
     python3 $pythonScriptsPath/associateDecalsBricksToFrames.py $referenceImagesForMosaic $imagesHdu $bricksIdentificationFile $brickDecalsAssociationFile $survey
 }
-export -f prepareDecalsDataForPhotometricCalibration
+export -f prepareSurveyDataForPhotometricCalibration
 
 # Photometric calibration functions
 # The function that is to be used (the 'public' function using OOP terminology)
@@ -1704,8 +1717,8 @@ computeCalibrationFactors() {
     numberOfFWHMForPhotometry=${12}
 
     mycatdir=$BDIR/my-catalog-halfmaxradius_it$iteration
-    methodToUse="noisechisel"
 
+    methodToUse="sextractor"
     echo -e "\n ${GREEN} ---Selecting stars and range for our data--- ${NOCOLOUR}"
     selectStarsAndSelectionRangeOurData $iteration $imagesForCalibration $mycatdir $methodToUse $tileSize
 
@@ -1969,6 +1982,7 @@ produceAstrometryCheckPlot() {
     astrometryTmpDir="./astrometryDiagnosisTmp"
     if ! [ -d $astrometryTmpDir ]; then mkdir $astrometryTmpDir; fi
     python3 $pythonScriptsPath/diagnosis_deltaRAdeltaDEC.py $matchCataloguesDir $output $pixelScale
+    rm -rf $astrometryTmpDir
 }
 export -f produceAstrometryCheckPlot
 
@@ -1991,30 +2005,36 @@ produceCalibrationCheckPlot() {
     for i in $myCatalogue_nonCalibrated/*.cat; do
         myFrame=$i
         frameNumber=$(echo "$i" | awk -F '[/]' '{print $(NF)}' | awk -F '[.]' '{print $(1)}' | awk -F '[_]' '{print $(NF)}')
-        referenceCatalogue=$referenceCatalogueDir/*_$frameNumber.*
 
-        myCalibratedFrame=$myFrames_calibrated/entirecamera_$frameNumber.fits
-        myNonCalibratedCatalogue=$myCatalogue_nonCalibrated/entirecamera_$frameNumber.fits*
-        fileWithMyApertureData=$aperturesForMyData_dir/range_entirecamera_$frameNumber*
+        # In the nominal resolution it takes sooo long for doing this plots. So only a set of frames are used for the
+        # calibration check
+        if [ "$frameNumber" -gt 5 ]; then
+            :
+        else
+            referenceCatalogue=$referenceCatalogueDir/*_$frameNumber.*
 
-        r_myData_pix_=$(awk 'NR==1 {printf $1}' $fileWithMyApertureData)
-        r_myData_pix=$(astarithmetic $r_myData_pix_ $numberOfFWHMToUse. x -q )
+            myCalibratedFrame=$myFrames_calibrated/entirecamera_$frameNumber.fits
+            myNonCalibratedCatalogue=$myCatalogue_nonCalibrated/entirecamera_$frameNumber.fits*
+            fileWithMyApertureData=$aperturesForMyData_dir/range_entirecamera_$frameNumber*
 
-        # raColumnName=RA
-        # decColumnName=DEC
-        # photometryOnImage_noisechisel -1 $tmpDir $myNonCalibratedCatalogue $myCalibratedFrame $r_myData_pix $tmpDir/$frameNumber.cat 22.5 \
-        #                                 $raColumnName $decColumnName
+            r_myData_pix_=$(awk 'NR==1 {printf $1}' $fileWithMyApertureData)
+            r_myData_pix=$(astarithmetic $r_myData_pix_ $numberOfFWHMToUse. x -q )
 
-        dataHdu=1
-        columnWithXCoordForOutDataPx=1 # These numbers come from how the catalogue of the matches stars is built. This is not very clear right now, should be improved
-        columnWithYCoordForOutDataPx=2
-        columnWithXCoordForOutDataWCS=3
-        columnWithYCoordForOutDataWCS=4
-        photometryOnImage_photutils -1 $tmpDir $myNonCalibratedCatalogue $myCalibratedFrame $r_myData_pix $tmpDir/$frameNumber.cat 22.5 $dataHdu \
-                                    $columnWithXCoordForOutDataPx $columnWithYCoordForOutDataPx $columnWithXCoordForOutDataWCS $columnWithYCoordForOutDataWCS
+            # raColumnName=RA
+            # decColumnName=DEC
+            # photometryOnImage_noisechisel -1 $tmpDir $myNonCalibratedCatalogue $myCalibratedFrame $r_myData_pix $tmpDir/$frameNumber.cat 22.5 \
+            #                                 $raColumnName $decColumnName
+            dataHdu=1
+            columnWithXCoordForOutDataPx=1 # These numbers come from how the catalogue of the matches stars is built. This is not very clear right now, should be improved
+            columnWithYCoordForOutDataPx=2
+            columnWithXCoordForOutDataWCS=3
+            columnWithYCoordForOutDataWCS=4
+            photometryOnImage_photutils -1 $tmpDir $myNonCalibratedCatalogue $myCalibratedFrame $r_myData_pix $tmpDir/$frameNumber.cat 22.5 $dataHdu \
+                                        $columnWithXCoordForOutDataPx $columnWithYCoordForOutDataPx $columnWithXCoordForOutDataWCS $columnWithYCoordForOutDataWCS
 
-        astmatch $referenceCatalogue --hdu=1 $tmpDir/$frameNumber.cat --hdu=1 --ccol1=RA,DEC --ccol2=RA,DEC --aperture=1/3600 --outcols=aMAGNITUDE,bMAGNITUDE -o$tmpDir/"$frameNumber"_matched.cat
-        rm $tmpDir/$frameNumber.cat
+            astmatch $referenceCatalogue --hdu=1 $tmpDir/$frameNumber.cat --hdu=1 --ccol1=RA,DEC --ccol2=RA,DEC --aperture=1/3600 --outcols=aMAGNITUDE,bMAGNITUDE -o$tmpDir/"$frameNumber"_matched.cat
+            rm $tmpDir/$frameNumber.cat
+        fi
     done
 
     python3 $pythonScriptsPath/diagnosis_magVsDeltaMag.py $tmpDir $output $outputDir $calibrationBrighLimit $calibrationFaintLimit $survey
